Adding data type to mz and intensities in load_from_msp

+++ matchms/importing/load_from_msp.py
@@ -48,8 +48,8 @@ def parse_msp_file(filename: str) -> Generator[dict, None, None]:
                     peakscount = 0
                     yield {
                         'params': (params),
-                        'm/z array': numpy.array(masses),
-                        'intensity array': numpy.array(intensities)
+                        'm/z array': numpy.array(masses, dtype=float),
+                        'intensity array': numpy.array(intensities, dtype=float)
                     }
 
                     params = {}
